fixed package structure of groups test
